+++ ChangeLog.rst
@@ -1,3 +1,9 @@
+Unreleased Changes
+==================
+
+* There is a new `s3qladm shrink-db` command that reclaims unused space in the
+  metadata database (both on disk and in the backend storage).
+
 S3QL 5.4.1 (2025-11-01)
 =======================